Fix for audio breaking after a certain time into the stream
After some time into the audio stream, the audio breaks and an error message is displayed in the console stating that the frame offset exceeds the total frames. 
``
caught RangeError: Failed to execute 'copyTo' on 'AudioData': Frame offset exceeds total frames (1024 >= 1024).
``
This is my proposed solution to the issue.  I have tested and verified that my fix works and the audio now runs smoothly without any break.

Hope it helps :)