Attempt some silliness
We can't send any objects until someone asks for them. The relay won't
ask until some subscriber asks, so we can just be that subscriber and
request our own tracks immediately. There may still be some issues with
this approach, but it's an idea at least.

The moq-js front end effectively ends up doing something similar because
it displays the round-tripped video back on the page after publishing.
Looking at the sequence of messages that sends is what inspired this hack.

We're not actually handing receiving those objects though, so this may
not be ideal for our case. As the comment indicates, it may be better to
just leave our track::Pulishers queuing up data ready to send once we
get a _real_ subscription request from some other subscriber on the
relay. We should even be able to take the same approach the relay itself
does and expire segments from those queues after some period of time to
prevent infinitely buffering data.