BREAKING: Use string paths instead of arrays. (#330)
moq-transport has an array of 32? binary parts. The idea is sound, but in practice it's really annoying to deal with an array of binary data.

moq-transfork slooowly weakened the requirement. It started as using utf-8 instead of binary data (which everybody does anyway). But when trying to integrate it with the

Additionally, prefix matching has always been supported but no suffix matching. I realized that if we allow one wildcard character, you can get both for free without any performance hit. More wildcard characters would start to hurt though. The ability to wildcards can still work with the old binary path but it's harder to encode in the message.

TODO: Escape wildcards.